Enlarged types

+++ Distributed_MatMul/include/MMMutilities.h
@@ -2,8 +2,8 @@
 #include <stdlib.h>
 #include "timings.h"
 
-void readBlockFromMatrix(double *block, double *matrix, uint nRows, uint nCols, uint N, uint startingCol);
+void readBlockFromMatrix(double *block, double *matrix, size_t nRows, size_t nCols, size_t N, size_t startingCol);
 
-void placeBlockInMatrix(double *block, double *matrix, uint nRows, uint nCols, uint N, uint startingCol);
+void placeBlockInMatrix(double *block, double *matrix, size_t nRows, size_t nCols, size_t N, size_t startingCol);
 
-void matMul(double *A, double *B, double *C, uint nRowsA, uint nColsARowsB, uint nColsB, uint startingCol, struct Timings* timings);
+void matMul(double *A, double *B, double *C, size_t nRowsA, size_t nColsARowsB, size_t nColsB, size_t startingCol, struct Timings* timings);
